requirements.txt
openai
pypdf
python-dotenv
pandas
argparse